+++ test-integration/edgar/integration_test_runner.clj
@@ -1,0 +1,7 @@
+(ns edgar.integration-test-runner
+  (:require [clojure.test :as t]
+            edgar.integration-test))
+
+(defn -main [& _]
+  (let [result (t/run-tests 'edgar.integration-test)]
+    (System/exit (if (= 0 (:fail result) (:error result)) 0 1))))